Rename colorsPalette util to namedColorsPalette
