+++ packages/grafana-ui/src/utils/namedColorsPalette.ts
