+++ public/app/plugins/panel/graph/data_processor.ts
@@ -1,8 +1,6 @@
 import _ from 'lodash';
-import { colors, GrafanaTheme } from '@grafana/ui';
-
+import { colors, GrafanaTheme, getColorFromHexRgbOrName } from '@grafana/ui';
 import TimeSeries from 'app/core/time_series2';
-import { getColorFromHexRgbOrName } from '@grafana/ui/src/utils/colorsPalette';
 import config from 'app/core/config';
 
 export class DataProcessor {
