+++ public/app/plugins/panel/heatmap/rendering.ts
@@ -8,8 +8,7 @@ import * as ticksUtils from 'app/core/utils/ticks';
 import { HeatmapTooltip } from './heatmap_tooltip';
 import { mergeZeroBuckets } from './heatmap_data_converter';
 import { getColorScale, getOpacityScale } from './color_scale';
-import { getColorFromHexRgbOrName } from '@grafana/ui/src/utils/colorsPalette';
-import { GrafanaTheme } from '@grafana/ui';
+import { GrafanaTheme, getColorFromHexRgbOrName } from '@grafana/ui';
 
 const MIN_CARD_SIZE = 1,
   CARD_PADDING = 1,
@@ -663,12 +662,10 @@ export class HeatmapRenderer {
 
   getCardColor(d) {
     if (this.panel.color.mode === 'opacity') {
-
       return getColorFromHexRgbOrName(
         this.panel.color.cardColor,
         contextSrv.user.lightTheme ? GrafanaTheme.Light : GrafanaTheme.Dark
       );
-
     } else {
       return this.colorScale(d.count);
     }
